Hand ioctl a request in the type both platforms give it
TIOCSWINSZ is not the same Swift type on either side of the portability
hedge. Darwin builds it out of the _IOW macro, which casts to unsigned
long, so it imports as UInt and matches ioctl's request parameter with
nothing to say about it. glibc defines it as a plain 0x5414, which imports
as Int32 — and the call then fails to compile with nothing
platform-specific visible in the line at all.

That is the same shape as the Darwin.write bug: a Darwin-ism with no #if
anywhere near it, which a macOS compiler cannot see and a macOS reviewer
has no reason to look at. It has been there since the initial commit and
is only failing now because §2.4's Linux job, added a day later, is what
can see it. Twice now that job has caught a real break rather than a
hypothetical future port, which is the argument for keeping it.

Converting explicitly is identity on Darwin and a widening on Linux, so
one line serves both. The comment is there because the conversion looks
gratuitous from macOS, which is how it came to be written without one.

+++ Sources/tetmuxCore/Core/PtyTransport.swift
@@ -328,7 +328,12 @@ public final class PtyTransport: @unchecked Sendable {
         lock.unlock()
         guard fd >= 0, cols > 0, rows > 0 else { return }
         var ws = winsize(ws_row: rows, ws_col: cols, ws_xpixel: 0, ws_ypixel: 0)
-        _ = ioctl(fd, TIOCSWINSZ, &ws)
+        // `ioctl` takes its request as a `UInt` on both platforms; the *constant* does not arrive as
+        // one on both. Darwin builds TIOCSWINSZ out of `_IOW`, which casts to `unsigned long`, so it
+        // imports as `UInt` and the bare call type-checks. glibc defines it as a plain `0x5414`,
+        // which imports as `Int32` — and the same line then fails to compile with nothing platform-
+        // specific visible in it. Converting explicitly is what makes one line serve both.
+        _ = ioctl(fd, UInt(TIOCSWINSZ), &ws)
     }
 
     /// Whether this transport still holds `pid` — i.e. nothing has waited on it yet, so the number
